Adds changes suggested in review.
